refactor: nesting for `constants` module
