+++ unicon_backend/models/__init__.py
@@ -1,6 +1,6 @@
 from sqlalchemy import create_engine
 
-from unicon_backend.helpers.constants import DATABASE_URL
+from unicon_backend.constants import DATABASE_URL
 from unicon_backend.models.base import Base
 from unicon_backend.models.contest import *  # noqa: F403
 from unicon_backend.models.user import User
